+++ src/onegov.swissvotes/HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 - Adds full text search support for attachments.
   [msom]
 
+- Adds voting booklet attachments.
+  [msom]
+
 - Makes the update, export and delete vote views to be indifferent to the filters.
   [msom]